req2test/requirements.txt
streamlit
pandas
python-docx
openai